pyproject.toml
[project]
name = "agentpdfx"
version = "0.1.0"
description = ""
authors = [
    {name = "Shubham Verma",email = "shubhamv8527@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4"
dependencies = [
    "agno (>=1.7.2,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "chromadb (>=1.0.15,<2.0.0)",
    "ollama (>=0.5.1,<0.6.0)",
    "pypdf (>=5.8.0,<6.0.0)",
    "mistralai (>=1.9.2,<2.0.0)",
    "pymilvus[model] (>=2.5.12,<3.0.0)",
]

[tool.poetry]
packages = [{include = "agentpdfx", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
